Add Mlps from final refinement to 3-spoke output
Full length Mlp1 and Mlp2 were included in the final refinement,
done after these RMFs were generated. For the sake of completeness,
add Mlp coordinates to the existing RMF file, replacing those that
were there already. This is done similarly to the addition to the
1-spoke output, except that two symmetry-related copies of
Mlp1 and Mlp2 are also added.